+++ src/vmmanager.rs
@@ -8,8 +8,8 @@ pub trait VMHypervisor{
     fn enter_state(&mut self, codata: &mut CoData, callsystem: & CallSystem) -> Result<(), NeutronError>;
     //note: hypervisors should own the relevant VM. Each execution (ie, sub-call etc) will produce a new VMHypervisor
     fn execute(&mut self, codata: &mut CoData) -> Result<VMResult, NeutronError>;
-    fn set_result(&mut self, result: u32);
-    fn set_error(&mut self, code: u32);
+    fn set_result(&mut self, result: u64);
+    fn set_error(&mut self, code: u64);
     /// Called when exiting the VM, should commit state etc
     fn exit_state(&mut self, codata: &mut CoData, callsystem: & CallSystem) -> Result<(), NeutronError>;
 }
